fix: extract normalizeToMonthlyEquivalents to own file to satisfy one-export-per-file lint rule (Story 90.1)
